+++ egui/src/any/mod.rs
@@ -49,9 +49,9 @@
 
 mod type_map;
 mod element;
-mod id_map;
+mod any_map;
 
 /// Same structs and traits, but also can be de/serialized under `persistence` feature.
 pub mod serializable;
 
-pub use self::{type_map::TypeMap, element::AnyMapTrait, id_map::AnyMap};
+pub use self::{type_map::TypeMap, element::AnyMapTrait, any_map::AnyMap};
