+++ egui/src/any/serializable/mod.rs
@@ -1,11 +1,11 @@
 mod type_map;
 #[cfg(feature = "persistence")]
 mod element;
-mod id_map;
+mod any_map;
 #[cfg(feature = "persistence")]
 mod type_id;
 
 #[cfg(not(feature = "persistence"))]
 use super::element;
 
-pub use self::{type_map::TypeMap, element::AnyMapTrait, id_map::AnyMap};
+pub use self::{type_map::TypeMap, element::AnyMapTrait, any_map::AnyMap};
